--- app/src/main/res/layout/activity_match.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<androidx.drawerlayout.widget.DrawerLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/drawer_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipChildren="false">
-
-        <RelativeLayout
-            android:id="@+id/button_container"
-            android:orientation="horizontal"
-            android:layout_width="match_parent"
-            android:layout_marginBottom="8dp"
-            android:layout_height="80dp"
-            android:layout_alignParentBottom="true"
-            android:clipChildren="false">
-
-            <com.google.android.material.floatingactionbutton.FloatingActionButton
-                android:id="@+id/skip_button"
-                android:contentDescription="Skip Button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerVertical="true"
-                android:layout_toStartOf="@+id/rewind_button"
-                android:src="@drawable/skip_red"
-                android:backgroundTint="@android:color/white"
-                app:fabSize="auto"
-                app:rippleColor="@color/purple_200"/>
-
-            <com.google.android.material.floatingactionbutton.FloatingActionButton
-                android:id="@+id/rewind_button"
-                android:contentDescription="Rewind Button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="32dp"
-                android:layout_marginEnd="32dp"
-                android:layout_centerInParent="true"
-                android:src="@drawable/rewind_blue"
-                android:backgroundTint="@android:color/white"
-                app:fabSize="mini"
-                app:rippleColor="@color/purple_200"/>
-
-            <com.google.android.material.floatingactionbutton.FloatingActionButton
-                android:id="@+id/like_button"
-                android:contentDescription="Like Button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerVertical="true"
-                android:layout_toEndOf="@+id/rewind_button"
-                android:src="@drawable/like_green"
-                android:backgroundTint="@android:color/white"
-                app:fabSize="auto"
-                app:rippleColor="@color/purple_200"/>
-
-        </RelativeLayout>
-
-        <com.yuyakaido.android.cardstackview.CardStackView
-            android:id="@+id/card_stack_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_above="@+id/button_container"
-            android:padding="4dp"
-            android:clipToPadding="false">
-        </com.yuyakaido.android.cardstackview.CardStackView>
-
-    </RelativeLayout>
-
-</androidx.drawerlayout.widget.DrawerLayout>
