+++ app/src/main/res/layout/fragment_match_page.xml
@@ -10,15 +10,69 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="300dp"
         android:orientation="vertical">
 
-        <Button
-            android:id="@+id/button_to_match"
+        <RelativeLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="100dp"
-            android:layout_marginRight="100dp"
-            android:text="@string/go_to_match_display"/>
+            android:layout_height="match_parent"
+            android:clipChildren="false">
+
+            <RelativeLayout
+                android:id="@+id/button_container"
+                android:layout_width="match_parent"
+                android:layout_height="80dp"
+                android:layout_alignParentBottom="true"
+                android:layout_marginBottom="8dp"
+                android:clipChildren="false"
+                android:orientation="horizontal">
+
+                <com.google.android.material.floatingactionbutton.FloatingActionButton
+                    android:id="@+id/skip_button"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_centerVertical="true"
+                    android:layout_toStartOf="@+id/rewind_button"
+                    android:backgroundTint="@android:color/white"
+                    android:contentDescription="Skip Button"
+                    android:src="@drawable/skip_red"
+                    app:fabSize="auto"
+                    app:rippleColor="@color/purple_200" />
+
+                <com.google.android.material.floatingactionbutton.FloatingActionButton
+                    android:id="@+id/rewind_button"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_centerInParent="true"
+                    android:layout_marginStart="32dp"
+                    android:layout_marginEnd="32dp"
+                    android:backgroundTint="@android:color/white"
+                    android:contentDescription="Rewind Button"
+                    android:src="@drawable/rewind_blue"
+                    app:fabSize="mini"
+                    app:rippleColor="@color/purple_200" />
+
+                <com.google.android.material.floatingactionbutton.FloatingActionButton
+                    android:id="@+id/like_button"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_centerVertical="true"
+                    android:layout_toEndOf="@+id/rewind_button"
+                    android:backgroundTint="@android:color/white"
+                    android:contentDescription="Like Button"
+                    android:src="@drawable/like_green"
+                    app:fabSize="auto"
+                    app:rippleColor="@color/purple_200" />
+
+            </RelativeLayout>
+
+            <com.yuyakaido.android.cardstackview.CardStackView
+                android:id="@+id/card_stack_view"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_above="@+id/button_container"
+                android:clipToPadding="false"
+                android:padding="4dp"></com.yuyakaido.android.cardstackview.CardStackView>
+
+        </RelativeLayout>
     </LinearLayout>
 </androidx.drawerlayout.widget.DrawerLayout>
